Separate SimpleMenu styles from MegaMenu styles
